Move ProgramTree to its own file
